+++ srcs/utils/unit.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   unit.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sharnvon <sharnvon@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/09 01:32:23 by sharnvon          #+#    #+#             */
+/*   Updated: 2023/03/09 03:34:12 by sharnvon         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minirt.h"
 
 t_color	set_color(int r, int g, int b)
@@ -10,7 +22,7 @@ t_color	set_color(int r, int g, int b)
 	return (color);
 }
 
-t_size set_size(double w, double h, double d)
+t_size	set_size(float w, float h, float d)
 {
 	t_size	size;
